Added Project file for QT-Creator

+++ src/ebevent.h
@@ -36,7 +36,7 @@
 #define EB_SIGNAL(name, args...) 	EBCpp::EBEvent<args> name
 #define EB_SLOT						/* for readability */
 
-#define EB_DEBUG(msg) std::cout << std::hex << "0x" << reinterpret_cast<int>(this) << std::dec << "(" << __PRETTY_FUNCTION__ << " " << __FILE__ << " @ " << __LINE__ << ") " << " " << msg << std::endl;
+#define EB_DEBUG(msg) std::cout << std::hex << "0x" << reinterpret_cast<long long>(this) << std::dec << "(" << __PRETTY_FUNCTION__ << " " << __FILE__ << " @ " << __LINE__ << ") " << " " << msg << std::endl;
 
 namespace EBCpp{
